Cleaned some code
BSD vs Linux.  This bash file is basically a simple set of terminal commands (obviously) and for whatever reason, BSD needed a few echo commands to move to the next line (comparable to you needing to press ENTER on your keyboard).  Turns out that was a bug with BSD and nothing to do with my code.  Which means it was safe to remove all the extra echo commands. -- Kept a few for display purposes.