initial release of collective Firefox auto install
The initial release of the collective Firefox automated installation process.  This file as only a bonus and was designed in mind for people who would have installed all editions (primarily developers) and as a means of saving them time.  The core goal of this development is to make installing native Mozilla Firefox easier and quicker.