permissions
The system default permissions for Linux for the path /opt/ is 755. As such, any new folder or file added below that path inherits those permissions.  This is inefficient for Mozilla's automatic update functionality. Since Mozilla Firefox needs permission to be able to write over its own files.  Originally, I had them set at 777 which was overkill. The happy medium appears to be 757.