Initial release
Just a basic document stating that this folder includes the standard release (64-bit) and where to local the obvious 32-bit releases if you need them.  As well as a reminder that the average user would not need to install both and to focus on how-to-install.txt for the simple install instructions.