addional release
Because someone actually contacted me and asked, "Where is the how-to-install.txt document?" Answer: In the primary directory. "Where is that?" Answer: The folder you extracted everything too.  "Oh. I do not see it in the 32bit folder".  *sigh* Not the primary folder but OK. -- Adding a copy of how-to-install.txt in both the 32bit folder and 64bit folder should make things easier.