linux is not apple
When using an Apple Computer I learned you do not need to add notations for the echo command.  I'm unsure if that is all Apple computers or my friend's computer has been tweaked.  Anyhow, when testing this script it worked just fine. On my Linux computer without the quotation marks ( broke the code.  I should have known better. But when you use a system you get use to that system.  

In short, the error was between developer and keyboard.