Redundancy
Originally I included a how-to-install.txt file in the root directly of the project, plus one in the 64-bit folder and 32-bit folder.  The idea of having 3 copies of the same documentation was to make this process more or less idiot-proof.  Since it would seem a few folks would dive right into either folder without looking at the 1st how-to-install.txt.  For simplicity, I hope that will no longer be needed since most folks should see the file "Install" and automatically be curious.  I can always re-add the extra set later if I must. lol