v2 uninstall user interface
v2 included the new user interface for installing Mozilla Firefox. It only seemed right we included that with the uninstaller as well.  You can still do it manually of course, but the interface makes it super newbie-friendly. 

Easter egg quote included all previous versions:  "That which is done, cannot be undone. Reinstalled, of course! But not undone."