v2 first dependency
I am actually proud of this release because I honestly feel v2 fulfills the promise truly making this more of an automated process for the basic home user. That including a user interface (as opposed to trying to pick the right file, which can still be done).  From here on out the user runs Install.sh and just follows the prompts (simple number selection). Even the documentation has become as simple as copy and paste (if they so choose). I don't think I could make it easier short of making the process as easy as setup.exe found in Windows and for Linux, I don't yet know how to make it that simple ... yet. lol

That all said, as noted in the file header  "This file to be used with Install.sh"  That is because the idea is the user runs Install.sh which allows them to pick between Firefox 64-bit or Firefox 32-bit.  Depending on which choice they make, the user interface will defer to either sub64.sh or sub32.sh  Making this the first mainstream internal dependency for the development.

Technically, firefox-all.sh (previously named, firefox-option.sh) was the first internal dependency. But I presume the vast majority of folks did not opt to install every edition of Mozilla Firefox, minus of course, developers and enthusiasts.