initial release v2
Firefox Automatic Install for Linux is now so simple I believe even a few who could be called computer illiterate could easily use it.  From the documentation being so straight forward you could basically copy & paste if you so choose and installing as simple as number selection.  

I'm proud of this release. It truly is an automatic experience (although it can still be used manually too, for those who choose)