case sensitive fix 2.0.1
The error between the developer and keyboard vs grammar.   English grammar would tell me Firefox being a name should be spelled with a capital "F" but the code called for a lowercase "f". This is purposely done for various reasons and I forgot to name the file correctly because my inner grammar police kicked in.