doc clean 2.0.1
Further cleaned the document.  The balance between not wanting the document to be a wall of a blog of text -vs- not be to lengthy -vs- clean layout to fit and read on a small screen -vs- informative -vs- perhaps a little education for true newbies... That's a lot to ask for in a small document. lol

I imagine the mission impossible theme song playing in the background anytime I think to edit this.  -- yes the developer has a sense of humor.