v2.5 Legacy Edition
The documentation was updated as Setup.sh now manages both the installation and uninstallation in a simple user interface. I call v2.5 "legacy edition" because this will be the last release with the "dos-like" interface. As we advance (v3 onward), future releases will focus on a graphic interface that can be used both via mouse and keyboard.