Icon generation
The icon generation is now handled independently of the primary install files. This of course is done in the background as the user will depend on Setup.sh to handle everything. The primary reason to separate the icon generation was only to make the primary install files easier to view from a developer's point of view.