dependency and icon defer
The new user interface requires the user to depend on Setup.sh for both install and uninstall. Additionally, icon generation is now been made to defer to its own independent script. This makes the primary install file easier to review from a developer's viewpoint (44 lines of code vs the original 300+).