defer icon
The primary change (beyond code cleanup) was to defer the icon generation code to its own independent file. This making managing and viewing the primary install file code easier (going from 300+ lines of code down to only 44 lines of code).  Additionally, the install setup is now dependent on Setup.sh, which will make the experience simpler for the end-user through the user interface.