2.5 documentation
Updated the ReadMe.md to reflect the new change of setup management. The user interface is all currently handled via Setup.sh for installation and uninstallation of Mozilla Firefox.  v2.5 is called the 'Legacy Edition' because it is the last planned version to depend on the old "dos-like" interface.  In the future, starting with v3, the focus will be developing and producing a user interface that will be more graphical, and the user can use both the mouse and keyboard to navigate easily.