purge
By default, if you choose to remove ALL editions of Mozilla Firefox, it should be assumed you wish to delete all the old cache files and the configuration and profile files. This update also specifically targets Firefox and not the Mozilla folder (directory), as someone recently reminded me Mozilla Thunderbird also stores info within the Mozilla folder. As someone who has been using webmail for the better part of a decade, this was an oversight on my part.