Icon name Firefox changed to Mozilla Firefox
This update renames the icon from simply "Firefox" to the name "Mozilla Firefox." 

It was (incorrectly) assumed that someone using "Firefox automatic install for Linux" would have first uninstalled the copy of Firefox, which may be pre-bundled in their Linux distro. The whole original point of this development was to "get away from" the customized copies often found within Linux distributions in favor of the pure stock experience offered directly from Mozilla.  

However, if someone were to keep both copies, the only "conflict" was the desktop (and menu) shortcuts were both named "Firefox." This technically had no direct impact on the install (or the system), only the icon shortcut. What could happen was should the distro copy update, it would often update the shortcut too. And as a result, you could unknowingly be redirected back to using the distro copy when using the shortcut.  The fix was as simple as being sure the icon shortcuts used different names. Since every distro I tested used the basic name "Firefox," I decided I could have picked between "Firefox Stable" or "Mozilla Firefox."  I decided to give credit where credit was due and selected "Mozilla Firefox."  The name "Mozilla Firefox" was also picked since it provided a cleaner and more professional naming scheme by using the full product name instead of calling it "Firefox stable."

The added benefit is you can now technically use both copies on the same system. This could be good for developers and anyone who wishes to compare the difference between the stock install -vs.- the pre-bundled copy often included in a Linux distro.

Obviously, anyone using the OEMs for distribution would not be including two copies of, Mozilla Firefox stable edition. I could have technically left the name for the OEM since it has nothing to do with Setup.sh designed for the end-user. But for uniformity of this development and not to underestimate or overestimate anyone, plus to give credit where credit is due, I have decided to apply this change to the OEM too, renaming the icon to "Mozilla Firefox."