Personal install
The purpose of this development (Firefox Automatic Install for Linux) is to help migrate away from customizations found within Linux distro copies of Mozilla Firefox. In favor of the pure default stock experience provided by the Mozilla Development. To that end, install Mozilla Firefox onto the computer running Linux for everyone using the computer and depend on Mozilla's and Linux's user profile management. Lastly, to accomplish it in such a way that doing so would be so newbie friendly that even a former (previous) Microsoft Windows user could reasonably achieve this goal. 

Over time, the development has expanded to include an "OEM" install folder designed for developers and distributions. However, to continue to do so using a silent install and without user prompts (for distribution purposes).

Today, I have not included a "personal" folder for individuals who are not system administrators who otherwise do not have root access. But for people who would still wish to experience the pure Mozilla stock experience. This, too, offers a silent install since the user would need only execute the corresponding file (for example, beta.sh would mean the beta edition of Mozilla Firefox).