For the love of shell
I love BASH. I cannot imagine not using BASH. But for maximized compatibility, I plan on updating the script to be more universal, where possible. At the moment, that means trying to get the code to work not only in BASH, but also ZSH, for example. And we start by replacing the old SheBang #!/bin/bash with #!bin/sh to tell the system to use their default script.