Expanded OEM Support - For the love of shell
BASH sell will always be the primary focus of this development. However, where possible, the code will be optimized to support alternative shells, for example, ZSH, so long as support is not limited or otherwise hindered in supporting BASH.  

I love BASH. I cannot imagine not using BASH. But for maximized compatibility, I plan on updating the script to be more universal, where possible. At the moment, that means trying to get the code to work not only in BASH, but also ZSH, for example. And we start by replacing the old SheBang #!/bin/bash with #!bin/sh to tell the system to use whatever default shell the user is using.