BASH -vs- ZSH - Options
"Options" is a special parameter. And while, although BASH has no problem with this, ZSH did. Renaming this to "menu" resolved the issue and allowed the code to run on either BASH or ZSH.

When possible, without taking away support for BASH, alternative code will be considered to support alternative shells.