Removed extra confirmation
Removing confirmation. The user instead is first given a warning before making their choice (from the sub menu). If they proceed, it was their choice to do so. They really do not need to be asked, "are you sure" a second time. Additionally, it cleans up and simplified the code.