file name change
I want to make learning easier. For example, while most people can easily figure out that fb32-icon meant "Firefox beta 32bit icon," I want it to be transparently obvious. And so for the would be new developer who decides to review the code, I want to make sure everything is easy to understand and by easy, I mean, I want it to be obviously transparent.