SUSE Linux
Once upon a time, there was a distro that, for reasons unknown, did not include libraries needed to extract a tar.bz2 file archive. In frustration, the developer of this project, had to add code that required anyone using openSUSE Linux to install the Bzip2 library as not to encounter an error in doing so. Then came the release of opneSUSE 15.4 which did at last resolve that nasty issue. -- finally. lol