Gecko Linux
Gecko Linux is a spin of openSUSE, but with all the cool non-free firmware and media codecs already included. The developer was smart enough to include Bzip2 libraries, but openSUSE was not, for reasons unknown. And so I need to tell openSUSE systems to install libraries needed to extract a simple tar.bz2 file. openSUSE finally fixed that with their Leap 15.4 release.  -- This note reflects both my relief, frustration, and a plug for Gecko Linux who did it right the first time. lol