SUSE dependency resolved
Of all the mainstream distros, only openSUSE, for reasons unknown, did not originally include libraries needed to extract a simple tar.bz2 file. So it was once needed to include code to add those, while also suppressing the error in non-openSUSE that zypper (their package manager) was not found. The code used was as follows:

zypper refresh --force 2> /dev/null
zypper install bzip2 2> /dev/null

But this code is no longer needed because, yay, openSUSE 15.4 includes the libraries.