SUSE Linux resolved
Bzip2 is now a standard package library, preinstalled and bundled with openSUSE 15.4. In a previous version, that was not the case, and you had no means to extract a simple tar.bz2 file.  Now that it's resolved, the extra code I used to add that to SUSE, while suppressing the error that zypper was not found, in non-SUSE distros, can finally go bye-bye.