openSUSE Linux Resolve
To my knowledge, every mainstream Linux distro includes the libraries needed to extract a simple tar.bz2 archive. For whatever reason, that was not so for openSUSE Linux. And for a time, I included some code to instruct openSUSE to install their Bzip2 library, while suppressing the error in all other distros, that zypper (their package manager) was not found. It's included now by default in openSUSE 15.4, so I was able to remove that extra code. I also don't need the documentation for the personal install, suggesting that someone make sure of that. That was a little extra note was included in the documentation, because the personal installer assumed you were NOT root or anyone who could apply that fix. But since it's now included in openSUSE by default, no one has to think about it. -- yay!