expanded suggestions
I added kate and leafpad as suggestions for desktop editors. I am primarily a KDE or XFCE user, myself. And in the many distros I use, those two seem to be common, by default.  I do not plan on extending my suggestion, as I know there are many editors.  But I did want to keep a few that I knew were user-friendly and commonly included in a pre-installation distro.