from echo to printf
The echo command, in my opinion, is easy and simple. But it is not always universally accepted and processed the same in different shells or different distros, which choose to add their own spin on things.  Honestly, sometimes, I wish Linux as a whole was more unified. That said, I am exploring the possibility of converting all those echo commands into "printf" commands.