simplified curl wget support
I rewrote the code in a more simplified manner instead of several lines of code (same results). The code will use either CURL or WGET, depending on what is available. And if neither one is available or should the download fail, it will redirect to an error screen. As a bonus, minus the desktop icon portion of the code (still working on that), this is now one more step closer to supporting both Fish Shell and Tcsh (Tenex C Shell), along with all POSIX-type shells, such as DASH, BASH, Zsh (Z-Shell), and Ksh (Korn Shell). So arguably, I have improved compatibility.