new error script
The development has always depended on WGET to download Mozilla Firefox. Recently I was surprised to learn that not every Linux distribution uses WGET (by default). To my surprise, some come with CURL out of the box instead. So I expanded the code install files to check for both and use whatever was available. However, in the unlikely event that someone has an edge case, not having either (seems highly unlikely), or if their download fails, I thought it good to have a simple error notice.