rewrite
This edition adds support for both CURL and WGET, as I became recently aware (and surprised) to learn some distributions do not include WGET. As someone who has always used openSUSE, Debian, Ubuntu, and alike this was news to me. The installer will not use whatever is available. However, if neither is available, it will correctly display a proper error message and safely exit without making any modifications. 

Additionally, this rewrite replaces the echo command with prinf for further compatibility. As explained to me, not all user shells or distros are equal and a simple command such as echo can prove problematic. However, printf seems to be more universal (I am paraphrasing).