printf vs echo
It has come to my attention that the simple command known as "echo" does not always function 100% the same in every user shell or distro. The command "printf" is however more universal (or so I am lead to believe). For this reason, I have rewritten the script to use instead of echo.