Execute Order 66
I added a purge option that not only uninstalls all 32-bit editions of Mozilla Firefox but also removes all cache and browser configuration for all users for all copies of Mozilla Firefox. This gives a nice clean slate (fresh start). -- I am a Star Wars fan too, so for a bit of humor, it is option #66 and when you select it, even says, "Per your request $USER. Execute Order 66."

Additionally, I have converted all the "echo" commands to use "printf" as I was recently informed not all shells or distros processes echo the same way. But printf is more universal.