replace echo with printf
Speed and uniformity -- As previously noted, I was surprised to learn that "echo" does not play universally in every shell or distro, but thankfully, "printf" does. And while I could have left it "as is" for the simple file creation for the icon shortcuts, changing 1 small line of code, makes the development complete the unification of the use of "printf."  Additionally, while I did not measure anything, it "seems" as though the file is quick with printf as opposed to echo (at least it did for me).