replace echo with printf
Improved compatibility and speed and all I had to do was stop using echo. Go figure. -- As it turns out, "echo" is not created equal in every distro or shell. For better compatibility and functionality, I replaced it with "printf"  -- Additionally, although I did not measure anything, it does "seem" to be faster to load too (but that's only an opinion).