replace echo with printf
As surprising as it is, "echo" is not universally built or functions depending on distro or shell. Universal compatibility is one of my general goals, and while I do not think it impacted the development, I like to keep the practice all the same.  Replacing "echo" with "printf" seems to better bring me closer to that goal.