Execute order 66
Can anyone tell that I am a Star Wars fan? 

In this release, I add a "purge" option for those who want a fresh, clean start. It includes a special caution notice outlining that this will purge all the cache and configuration files for all copies of Mozilla Firefox on your computer. This is ideal for developers or anyone else needing a clean profile and a fresh start.  I made this option a double-digit command so no one could mistakenly press a single number by mistake. I picked 66 for obvious reasons and found it fitting.

Additionally, this release also replaces "echo" with "printf." Much to my surprise, I learned echo is not universal depending on distro and shell. It has been part of my driving goal to make the script as universally compatible as possible.  Using "prinf" further along that objective.