CURL support with WGET
My distro of choice is usually openSUSE or Debian or alike, and WGET was provided by default. I recall testing the top 5-10 on distrowatch.com, and WGET was always standard. Imagine my surprise when I learned that may technically be the exception and not the norm. -- This release will not use whichever is available (WGET or CURL). And if either fails, display a proper error message. 

Additionally, I have replaced "echo" with "printf" as I was made aware that echo is not universal in how it functions or performs in every distro or shell. Whenever possible, it has been my goal to maximize compatibility, and to that end, I am retiring echo in replacing it with printf.