Curl support with Wget
I was recently surprised to learn that WGET, is not commonly included in most distros. Being that my distro of choice would normally be OpenSUSE, Debian, Ubuntu, and alike. Stranger still, I recall testing the top 5-10 distros listed on distrowatch.com to verify the functionality of my script, and in all instances, WGET was included.  In any case, it will now use whichever is available (CURL or WGET), and if something fails, it will display a proper error message.