Linux fragmentation
I incorrectly, based upon continuous experience, assumed WGET was standard in all distros. My primary distro of choice is openSUSE or Debian, and alike. Surprise, it's not, and as it turns out CURL is more common. It is ironic since I always needed to install, CURL.  In any case, the setup will now automatically check for both WGET and CURL. If neither is available or if the download fails, you will be greeted with a friendly error message.