why printf "\n" after done
I forgot to add printf "\n" when I was replacing echo -n.  --- So, why do I add this?

In a few rare instances (edge case), when the script hit "done" it hangs. Because for whatever reason "done' did not mean just that portion of the script, it registered as "done" as if "finished" but without any exit. This caused it to hang and not proceed further.  Adding a useless "echo -n" resolved it in those rare instances.