Curl and Wget
If you're an openSUSE or Debian user or similar, you may be surprised to learn WGET is not standard in most distributions.  When making this script, I had tested more than my fair share off of distrowatch.com and WGET had been ironically included in all of them.  I was surprised to learn that this is not the norm.  Going forward, the setup will not pick between WGET and CURL, ensuring compatibility. And should that fail, you will be greeted with a friendly error message.