echo vs printf
Echo has always seemed so simple, so easy, and so dependable. Imagine my surprise when it became known to me, that echo is not always universally the same depending on shell or distro. While it did not impact this development, specifically, the goal has always been to maximize compatibility. To that end, printf which is more universally the same will be used to replace echo, going forward.