2.9.0.3
I extended download support because, as it turns out, I was living in a bubble. All my distributions of choice include WGET by default. So I incorrectly assumed that was the norm. Ironically, I had even tested many distros off of distrowatch.com, and they reassured my belief that WGET was the norm. It's not as CURL is widely more common. The development now will automatically search to use either WGET or CURL. If neither is available, you will be greeted by a friendly error message.

Another new feature I added to the uninstall menu is "purge."  This is ideal for developers or power users who need a fresh new start.

Lastly, it came to my attention and understanding that "echo" is not as nearly widely universal depending on what shell or distribution you are running. "Printf," however, is, so the whole development, including the personal installer and OEM, has retired "echo" in favor of "printf."  For the user interface,  a whole new rewrite.