CURL support added
I was surprised to learn that "WGET" is not usually included in most distributions. My distribution of choice is often openSUSE or Debian and alike. I even recall testing more than a dozen random distros from distowatch.com, but they included "WGET." In short, I was living in a bubble. -- The script now will search for "WGET" and "CURL."  A friendly error message will display if neither is available.