path name change
from 32bit-uninstallers to 32bit_uninstallers.  Why? I encountered a weird bug that caused this to register as a flag. It was an edge case since it only impacted a single distro, and they fixed it, but I see no harm in making sure this never happens in the future.   Lastly, I think it makes it easier to read too, so, why not?!