partial fish shell support
The OEM and personal installer will now also work with Fish Shell in addition to all the previous shells it already supports (Posix, bash, dash, ksh, zsh).  While the normal install performed through Setup.sh is not yet compatible (the menu needs to be rewritten), the code change for the actual install files is applied. -- Still, that's impressive considering now only 4 files are all that are left for full backward compatibility.